+++ public/install/data/mall_goods.sql
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS `np_mall_goods` (
   `is_com` tinyint(1) unsigned NOT NULL DEFAULT '0' COMMENT '推荐',
   `is_top` tinyint(1) unsigned NOT NULL DEFAULT '0' COMMENT '置顶',
   `is_hot` tinyint(1) unsigned NOT NULL DEFAULT '0' COMMENT '最热',
-  `sort` int(11) unsigned NOT NULL DEFAULT '0' COMMENT '排序',
+  `sort` smallint(6) unsigned NOT NULL DEFAULT '0' COMMENT '排序',
   `hits` int(11) unsigned NOT NULL DEFAULT '0' COMMENT '点击量',
   `comment_count` int(11) unsigned NOT NULL DEFAULT '0' COMMENT '评论量',
   `update_time` int(11) unsigned NOT NULL DEFAULT '0' COMMENT '修改时间',
